move dmsg init logic to internal/cli
